refactor(storage-manager): regroup code within same file
The methods `create_and_start_storage` and `start_storage` were in two
separate files -- the former calling the latter.

This commit regroups them in the same file.

* plugins/zenoh-plugin-storage-manager/src/backends_mgt.rs: deleted the
  file, exporting its content to `storages_mgt.rs`.
* plugins/zenoh-plugin-storage-manager/src/lib.rs: fixed the imports.
* plugins/zenoh-plugin-storage-manager/src/replica/mod.rs: fixed the
  imports.
* plugins/zenoh-plugin-storage-manager/src/replica/storage.rs: fixed the
  imports.
* plugins/zenoh-plugin-storage-manager/src/storages_mgt.rs: imported the
  code contained in `backends_mgt.rs`.
